+++ src/care_odoo/resources/product_product/spec.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from enum import Enum
 
 from pydantic import BaseModel
@@ -7,7 +8,7 @@ from care_odoo.resources.product_category.spec import CategoryData
 
 class TaxData(BaseModel):
     tax_name: str
-    tax_percentage: float
+    tax_percentage: Decimal
 
 
 class ProductStatus(str, Enum):
@@ -19,8 +20,8 @@ class ProductStatus(str, Enum):
 class ProductData(BaseModel):
     product_name: str
     x_care_id: str
-    cost: float
-    mrp: float
+    cost: Decimal
+    mrp: Decimal
     category: CategoryData
     taxes: list[TaxData] | None = None
     hsn: str | None = None
